chore(tests): refactor test utilities across files
