+++ src/tests/utils/constants.ts
@@ -1,0 +1,7 @@
+/**
+ * Just some hard coded strings for the tests.
+ * Do not actually use values like this in an actual project/application.
+ */
+export const COOKIE_SECRET = "some other secret, do not make them the same";
+export const HEADER_KEY = "x-csrf-token";
+export const TEST_TOKEN = "test token";
